Honor the boot timeout argument
dotest.py accepts a "--boot_time" argument, which at the moment is
only used by "BootTestCase".

For the other test suites, the global test timeout, passed through
"--timeout" is used when waiting for emulator instances to boot.

This change ensures that the "--boot_time" value will be used when
waiting for instances to boot. Having a separate timeout is really
usefull, especially when dealing with hanging instances.